Added support for tyre data files which are basically INI files with the use of a $ character for comments.
This has been implemented by adding an input argument strict to the C functions that parse the string data into numerical values.  When strict=1 the behaviour is the same as before for an INI file whereby any characters after the value will cause an error to be returned.  When strict=0 the characters after the value are ignored and the value returned.

A sample.tir is included to support the test case.